% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateThreshVsPerfData.R
\name{generateThreshVsPerfData}
\alias{generateThreshVsPerfData}
\title{Generate threshold vs. performance(s) for 2-class classification.}
\usage{
generateThreshVsPerfData(e, gridsize = 100L)
}
\arguments{
\item{e}{Container object for a machine learning experiment.}

\item{gridsize}{Grid resolution for x-axis (threshold). Default is 100.}
}
\description{
Generates data on threshold vs. performance(s) for 2-class classification that can be used for plotting.
}
\examples{
1
}
