% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateThreshVsPerfData.R
\name{generateThreshVsPerfDataInBenchmark}
\alias{generateThreshVsPerfDataInBenchmark}
\title{Generate threshold vs. performance(s) for 2-class classification in Benchmark object.}
\usage{
generateThreshVsPerfDataInBenchmark(bmr, gridsize = 100L)
}
\arguments{
\item{bmr}{Benchmark object.}

\item{gridsize}{Grid resolution for x-axis (threshold). Default is 100.}
}
\description{
Generates data on threshold vs. performance(s) for 2-class classification that can be used for plotting.
}
\examples{
1
}
