% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3vis.R
\docType{package}
\name{mlr3vis-package}
\alias{mlr3vis}
\alias{mlr3vis-package}
\title{mlr3vis: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
