% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrVis.R
\name{plotBMRBoxplots}
\alias{plotBMRBoxplots}
\title{Create box, violin or dot plots for a BenchmarkResult.}
\usage{
plotBMRBoxplots(bmr, style = "box", xVar = "learner_id",
  facet_x = "task_id", facet_y = "measure")
}
\arguments{
\item{bmr}{the BenchmarkResult object}

\item{style}{Type of plot, can be \dQuote{box} for a boxplot, \dQuote{violin} for a violin plot,
or \dQuote{dot} for a dot plot. Default is \dQuote{box}.}

\item{xVar}{variable for X axis, can be learner_id, task_id, measure}

\item{facet_x}{variable for facetting in columns, can be learner_id, task_id, measure, or NULL}

\item{facet_y}{variable for facetting in rows, can be learner_id, task_id, measure, or NULL}
}
\description{
Plots box, violin or dot plots for all `measure` across all iterations
of the resampling strategy in all `learner_id`, faceted by the `task.id` and `measure`.
}
