% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotThreshVsPerf.R
\name{plotThreshVsPerf}
\alias{plotThreshVsPerf}
\title{Plot threshold vs. performance(s) for 2-class classification using ggplot2.}
\usage{
plotThreshVsPerf(performanceByThreshold)
}
\arguments{
\item{performanceByThreshold}{object (data.table) generated by generateThreshVsPerfData.}
}
\description{
Plots threshold vs. performance(s) data that has been generated with generateThreshVsPerfData.
}
\examples{
1
}
